Changes after code review part 1
