+++ src/server/routes/controllers/logInUser.js
@@ -1,9 +1,10 @@
+const logInUtilities = require('./utilities/logInUtilities')
 const strategyUtils = require('../../passport/strategyUtils')
 
 function logIn(passport = require('passport')) {
 	return (req, res, next) => {
-		const loginStrategyHandler = strategyUtils.strategyHandlers(req, res, next).loginStrategyHandler
-		const logInByPassport = passport.authenticate(strategyUtils.STRATEGY_NAME, loginStrategyHandler) 
+		const proceedLogIn = logInUtilities(req, res, next).proceedLogIn
+		const logInByPassport = passport.authenticate(strategyUtils.STRATEGY_NAME, proceedLogIn) 
 		logInByPassport(req, res, next)
 	}
 }
